backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.4
catboost==1.2.8
scikit-learn==1.6.1
gunicorn==23.0.0
